fix: Fix app never updating because of stale service worker
Updating to create-react-app v4 changed how service workers work, and
service-worker.js was no longer being included in the build. This meant
that the previous service worker never updated, which resulted in new
code never being served because existing visitors to the site received
code from an old service worker that was cached locally. Adding an
empty service-worker.js file will force the browser to update the
service-worker.js, which results in getting the new code, which then
unregisters the service worker, but only after the user has closed the
tab and re-opened it. Service workers are tricky! Be careful!